[BM_{OSAL|HAL}] Add HAL layer to remove tight coupling on STM chip
