+++ src/lib/bm_hal/bm_hal.h
@@ -1,0 +1,15 @@
+#ifndef BM_HAL_H
+#define BM_HAL_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void bm_hal_wd_reload_counter(void);
+void bm_hal_wd_log(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // BM_HAL_H
