+++ src/lib/bm_osal/bm_osal_freertos.c
@@ -48,8 +48,8 @@ bool bm_osal_queue_delete(bm_osal_queue_t queue) {
 }
 
 // Task
-bool bm_osal_task_create(bm_osal_task_func_t task_func, const char* name, uint16_t stack_size, void* arg,  uint32_t priority, bm_task_handle_t task_handle) {
-    return (xTaskCreate(task_func, name, stack_size, arg, priority, task_handle) == pdPASS);
+bool bm_osal_task_create(bm_osal_task_func_t task_func, const char* name, uint16_t stack_size, void* arg,  uint32_t priority, bm_osal_task_handle_t* task_handle) {
+    return (xTaskCreate(task_func, name, stack_size, arg, priority, (TaskHandle_t *)task_handle) == pdPASS);
 }
 
 bool bm_osal_task_delete(void) {
@@ -86,26 +86,4 @@ bool bm_osal_timer_delete(bm_osal_timer_t timer) {
 // Delay
 void bm_osal_delay_ms(uint32_t ms) {
     vTaskDelay(pdMS_TO_TICKS(ms));
-}
-
-void bm_hal_wd_reload_counter() {
-    LL_IWDG_ReloadCounter(IWDG);
-}
-
-// void* bm_hal_wd_get_handle(void) {
-//     return IWDG;
-// }
-
-void IWDG_IRQHandler(void) {
-  LL_IWDG_EnableWriteAccess(IWDG);
-  while (LL_IWDG_IsReady(IWDG) != 1)
-  {
-  }
-  LL_IWDG_SetPrescaler(IWDG, LL_IWDG_PRESCALER_1024);
-  LL_IWDG_ReloadCounter(IWDG);
-  MEMFAULT_ASSERT_EXTRA_AND_REASON(0, kMfltRebootReason_HardwareWatchdog);
-}
-
-void bm_hal_wd_log() {
-//   memfault_software_watchdog_feed();
 }
